Added vehicle image force deleting check
Now vehicle images are only deleted if you are completely deleting a vehicle, and not if you are discontinuing it

+++ app/Listeners/VehicleDeleting.php
@@ -17,6 +17,8 @@ class VehicleDeleting
     public function handle(VehicleDeletingEvent $event)
     {
         $vehicle = $event->vehicle;
-        $vehicle->deleteImages();
+        if ($vehicle->isForceDeleting()) {
+            $vehicle->deleteImages();
+        }
     }
 }
